# config/swagger-output.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "assignment-5-customer-data",
    "version": "1.0.0",
    "description": "API for managing customers",
    "contact": {}
  },
  "servers": [],
  "paths": {
...
    "/customers/{customerId}/orders": {
      "post": {
        "summary": "Create a new order for a specific customer",
        "description": "Create a new order and associate it with a specific customer using the customer's ID.",
        "operationId": "createOrder",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer for whom the order is created.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateOrderRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successfully created a new order for the customer.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Customer not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all orders for a specific customer",
        "description": "Retrieve all orders associated with a specific customer.",
        "operationId": "getAllOrder",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer whose orders are being retrieved.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all orders for the customer.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Order"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/customers/{customerId}/orders/{orderId}": {
      "get": {
        "summary": "Get an order by ID",
        "description": "Retrieve a specific order by its ID, including the associated customer details.",
        "operationId": "getOrderById",
        "tags": ["Customers"],
        "parameters": [
          {
            "name": "customerId",
            "in": "path",
            "required": true,
            "description": "The ID of the customer.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          },
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "The ID of the order to retrieve.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved the order by ID.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Order"
                }
              }
            }
          },
          "404": {
            "description": "Order not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/orders/{orderId}": {
      "put": {
        "summary": "Update an existing order",
        "description": "Update details of an existing order by its ID.",
        "operationId": "updateOrder",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "The ID of the order to update.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ord_date": {
                    "type": "string",
                    "format": "date",
                    "description": "The date of the order",
                    "example": "2023-10-15"
                  },
                  "ord_status": {
                    "type": "string",
                    "description": "The status of the order",
                    "example": "Completed"
                  },
                  "ord_totalAmount": {
                    "type": "number",
                    "format": "float",
                    "description": "The total amount of the order",
                    "example": 150.75
                  }
                },
                "required": ["ord_date", "ord_status", "ord_totalAmount"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successfully updated the order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success update order"
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "previousDataOrder": {
                          "$ref": "#/components/schemas/Order"
                        },
                        "updatedOrder": {
                          "type": "object",
                          "properties": {
                            "ord_date": {
                              "type": "string",
                              "format": "date",
                              "example": "2023-10-15"
                            },
                            "ord_status": {
                              "type": "string",
                              "example": "Completed"
                            },
                            "ord_totalAmount": {
                              "type": "number",
                              "example": 150.75
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - all fields must be filled.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "All fields must be filled"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete an order by ID",
        "description": "Delete a specific order by its ID.",
        "operationId": "deleteOrder",
        "tags": ["Orders"],
        "parameters": [
          {
            "name": "orderId",
            "in": "path",
            "required": true,
            "description": "The ID of the order to delete.",
            "schema": {
              "type": "integer",
              "example": 1
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully deleted the order.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Success delete order"
                    },
                    "data": {
                      "$ref": "#/components/schemas/Order"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Order not found.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Order not found"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Internal server error"
                    }
                  }
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Order": {
        "type": "object",
        "properties": {
          "ord_id": {
            "type": "integer",
            "description": "The unique identifier for the order",
            "example": 1
          },
          "ord_date": {
            "type": "string",
            "format": "date",
            "description": "The date when the order was created",
            "example": "2023-10-15"
          },
          "ord_status": {
            "type": "string",
            "description": "The current status of the order",
            "example": "Completed"
          },
          "ord_totalAmount": {
            "type": "number",
            "format": "float",
            "description": "The total amount for the order",
            "example": 150.75
          },
          "ord_createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was created",
            "example": "2023-10-15T14:48:00.000Z"
          },
          "ord_updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the order was last updated",
            "example": "2023-10-16T10:15:00.000Z"
          }
        },
        "required": ["ord_id", "ord_date", "ord_status", "ord_totalAmount"]
      },
...
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Error message indicating the issue",
            "example": "Order not found"
          }
        }
      },
      "CreateOrderRequest": {
        "type": "object",
        "properties": {
          "ord_date": {
            "type": "string",
            "format": "date",
            "example": "2023-10-15"
          },
          "ord_status": { "type": "string", "example": "Pending" },
          "ord_totalAmount": {
            "type": "number",
            "format": "float",
            "example": 200.5
          }
        },
        "required": ["ord_date", "ord_status", "ord_totalAmount"]
      }
    }
  }
}